XP-641 Image cropping - Add crop button
- updated icomoon with new icon
- it seems that icomoon icon names/values were updated so had to fix occurrences
- added crop icon and crop edit toolbar to focal editor
- renamed focal editor to image editor
