+++ modules/admin-ui/src/main/resources/web/admin/common/icons/icons.less
@@ -55,159 +55,73 @@
   }
 }
 
-.icon-checkmark:extend(.icon) {
-  &:before {
-    content: "\e370";
-  }
-}
-
-.icon-arrow-down4 {
-  &:before {
-    .icon();
-    content: "\e3c5";
-  }
-}
-
-.icon-arrow-up4 {
-  &:before {
-    .icon();
-    content: "\e3c1";
-  }
-}
-
-.icon-menu3 {
-  &:before {
-    .icon();
-    content: "\e2ac";
-  }
-}
-
-.icon-close {
+.icon-center_focus_strong {
   &:extend(.icon);
   &:before {
-    content: "\e36b";
+    content: "\e149";
   }
 }
 
-.icon-arrow-up-thin {
+.icon-crop {
   &:extend(.icon);
   &:before {
-    content: "\e410";
+    content: "\e153";
   }
 }
 
-.icon-arrow-right-thin {
+.icon-newspaper {
   &:extend(.icon);
-  &:before {
-    content: "\e411";
-  }
-}
-
-.icon-arrow-down-thin {
-  &:extend(.icon);
-  &:before {
-    content: "\e412";
-  }
-}
-
-.icon-arrow-left-thin {
-  &:extend(.icon);
-  &:before {
-    content: "\e413";
-  }
-}
-
-.icon-arrow-down {
-  &:before {
-    .icon();
-    content: "\e3b2";
-  }
-}
-
-.icon-arrow-right2:extend(.icon) {
-  &:before {
-    content: "\e3b1";
-  }
-}
-
-.icon-arrow-right3:extend(.icon) {
-  &:before {
-    content: "\e3b5";
-  }
-}
-
-.icon-arrow-right6 {
-  &:extend(.icon);
-  &:before {
-    content: "\e3cb";
-  }
-}
-
-.icon-grid:extend(.icon) {
-  &:before {
-    content: "\e2a0";
-  }
-}
-
-.icon-grid4 {
-  &:extend(.icon);
-  &:before {
-    content: "\e2a3";
-  }
-}
-
-.icon-image:extend(.icon) {
-  &:before {
-    content: "\e02b";
-  }
-}
-
-.icon-blocked:extend(.icon) {
-  &:before {
-    content: "\e364";
-  }
-}
-
-.icon-type:extend(.icon) {
-  &:before {
-    content: "\e449";
-  }
-}
-
-.icon-close2 {
-  &:extend(.icon);
-  &:before {
-    content: "\e36c";
-  }
-}
-
-.icon-puzzle:extend(.icon) {
-  &:before {
-    content: "\e26e";
-  }
-}
-
-.icon-earth:extend(.icon) {
-  &:before {
-    content: "\e2c7";
-  }
-}
-
-.icon-insert-template:extend(.icon) {
-  &:before {
-    content: "\e467";
-  }
-}
-
-.icon-newspaper:extend(.icon) {
   &:before {
     content: "\e00c";
   }
 }
 
-.icon-wand:extend(.icon) {
+.icon-image {
+  &:extend(.icon);
   &:before {
-    content: "\e1ec";
+    content: "\e02b";
+  }
+}
+
+.icon-file {
+  &:extend(.icon);
+  &:before {
+    content: "\e077";
+  }
+}
+
+.icon-file-upload {
+  &:extend(.icon);
+  &:before {
+    content: "\e08a";
+  }
+}
+
+.icon-folder {
+  &:extend(.icon);
+  &:before {
+    content: "\e098";
+  }
+}
+
+.icon-address-book {
+  &:extend(.icon);
+  &:before {
+    content: "\e0e8";
+  }
+}
+
+.icon-clock {
+  &:extend(.icon);
+  &:before {
+    content: "\e101";
+  }
+}
+
+.icon-calendar {
+  &:extend(.icon);
+  &:before {
+    content: "\e115";
   }
 }
 
@@ -215,20 +129,6 @@
   &:extend(.icon);
   &:before {
     content: "\e138";
-  }
-}
-
-.icon-target3 {
-  &:extend(.icon);
-  &:before {
-    content: "\e27b";
-  }
-}
-
-.icon-stats {
-  &:extend(.icon);
-  &:before {
-    content: "\e1fe";
   }
 }
 
@@ -246,59 +146,45 @@
   }
 }
 
-.icon-address-book {
+.icon-quotes-left {
   &:extend(.icon);
   &:before {
-    content: "\e0e8";
+    content: "\e19f";
   }
 }
 
-.icon-arrow-down2 {
+.icon-search {
   &:extend(.icon);
   &:before {
-    content: "\e3b6";
+    content: "\e1b7";
   }
 }
 
-.icon-loop5 {
+.icon-wand {
   &:extend(.icon);
   &:before {
-    content: "\e3a9";
+    content: "\e1ec";
   }
 }
 
-.icon-search3 {
+.icon-stats {
   &:extend(.icon);
   &:before {
-    content: "\e1ba";
+    content: "\e1fe";
   }
 }
 
-.icon-calendar4 {
+.icon-puzzle {
   &:extend(.icon);
   &:before {
-    content: "\e115";
+    content: "\e26e";
   }
 }
 
-.icon-clock {
+.icon-target {
   &:extend(.icon);
   &:before {
-    content: "\e101";
-  }
-}
-
-.icon-file {
-  &:extend(.icon);
-  &:before {
-    content: "\e077";
-  }
-}
-
-.icon-folder {
-  &:extend(.icon);
-  &:before {
-    content: "\e098";
+    content: "\e27b";
   }
 }
 
@@ -309,38 +195,10 @@
   }
 }
 
-.icon-file-upload2 {
+.icon-list {
   &:extend(.icon);
   &:before {
-    content: "\e08a";
-  }
-}
-
-.icon-paragraph-left {
-  &:extend(.icon);
-  &:before {
-    content: "\e46b";
-  }
-}
-
-.icon-paragraph-center {
-  &:extend(.icon);
-  &:before {
-    content: "\e46c";
-  }
-}
-
-.icon-paragraph-right {
-  &:extend(.icon);
-  &:before {
-    content: "\e46d";
-  }
-}
-
-.icon-paragraph-justify {
-  &:extend(.icon);
-  &:before {
-    content: "\e46e";
+    content: "\e299";
   }
 }
 
@@ -351,10 +209,31 @@
   }
 }
 
-.icon-list {
+.icon-grid {
   &:extend(.icon);
   &:before {
-    content: "\e299";
+    content: "\e2a0";
+  }
+}
+
+.icon-grid2 {
+  &:extend(.icon);
+  &:before {
+    content: "\e2a3";
+  }
+}
+
+.icon-menu {
+  &:extend(.icon);
+  &:before {
+    content: "\e2ac";
+  }
+}
+
+.icon-earth {
+  &:extend(.icon);
+  &:before {
+    content: "\e2c7";
   }
 }
 
@@ -365,10 +244,129 @@
   }
 }
 
-.icon-quotes-left {
+.icon-blocked {
   &:extend(.icon);
   &:before {
-    content: "\e19f";
+    content: "\e364";
+  }
+}
+
+.icon-close {
+  &:extend(.icon);
+  &:before {
+    content: "\e36b";
+  }
+}
+
+.icon-close2 {
+  &:extend(.icon);
+  &:before {
+    content: "\e36c";
+  }
+}
+
+.icon-checkmark {
+  &:extend(.icon);
+  &:before {
+    content: "\e370";
+  }
+}
+
+.icon-loop {
+  &:extend(.icon);
+  &:before {
+    content: "\e3a9";
+  }
+}
+
+.icon-arrow-right {
+  &:extend(.icon);
+  &:before {
+    content: "\e3b1";
+  }
+}
+
+.icon-arrow-down {
+  &:extend(.icon);
+  &:before {
+    content: "\e3b2";
+  }
+}
+
+.icon-arrow-right2 {
+  &:extend(.icon);
+  &:before {
+    content: "\e3b5";
+  }
+}
+
+.icon-arrow-down2 {
+  &:extend(.icon);
+  &:before {
+    content: "\e3b6";
+  }
+}
+
+.icon-arrow-up {
+  &:extend(.icon);
+  &:before {
+    content: "\e3c1";
+  }
+}
+
+.icon-arrow-down3 {
+  &:extend(.icon);
+  &:before {
+    content: "\e3c5";
+  }
+}
+
+.icon-arrow-right3 {
+  &:extend(.icon);
+  &:before {
+    content: "\e3cb";
+  }
+}
+
+.icon-arrow-up2 {
+  &:extend(.icon);
+  &:before {
+    content: "\e410";
+  }
+}
+
+.icon-arrow-right4 {
+  &:extend(.icon);
+  &:before {
+    content: "\e411";
+  }
+}
+
+.icon-arrow-down4 {
+  &:extend(.icon);
+  &:before {
+    content: "\e412";
+  }
+}
+
+.icon-arrow-left {
+  &:extend(.icon);
+  &:before {
+    content: "\e413";
+  }
+}
+
+.icon-font-size {
+  &:extend(.icon);
+  &:before {
+    content: "\e448";
+  }
+}
+
+.icon-type {
+  &:extend(.icon);
+  &:before {
+    content: "\e449";
   }
 }
 
@@ -400,16 +398,37 @@
   }
 }
 
-.icon-font-size {
+.icon-insert-template {
   &:extend(.icon);
   &:before {
-    content: "\e448";
+    content: "\e467";
   }
 }
 
-.icon-center-focus-strong {
+.icon-paragraph-left {
   &:extend(.icon);
   &:before {
-    content: "\e600";
+    content: "\e46b";
+  }
+}
+
+.icon-paragraph-center {
+  &:extend(.icon);
+  &:before {
+    content: "\e46c";
+  }
+}
+
+.icon-paragraph-right {
+  &:extend(.icon);
+  &:before {
+    content: "\e46d";
+  }
+}
+
+.icon-paragraph-justify {
+  &:extend(.icon);
+  &:before {
+    content: "\e46e";
   }
 }
