+++ modules/admin-ui/src/main/resources/web/admin/common/js/ui/image/_module.ts
@@ -1,2 +1,2 @@
 ///<reference path='ImageCanvas.ts' />
-///<reference path='FocalEditor.ts' />
+///<reference path='ImageEditor.ts' />
