+++ modules/admin-ui/src/main/resources/web/admin/common/styles/api/app/appbar.less
@@ -15,7 +15,7 @@
 
   .launcher-button {
     &:extend(.icon);
-    .icon-grid4;
+    .icon-grid2;
 
     .flex(0 0 30px);
     .order(3);
