+++ modules/admin-ui/src/main/resources/web/admin/common/styles/api/ui/cycle-button.less
@@ -56,7 +56,7 @@
       .icon-arrow-down();
     }
 
-    .icon-loop5();
+    .icon-loop();
   }
 
 }
