+++ modules/admin-ui/src/main/resources/web/admin/common/styles/api/ui/time/date-picker.less
@@ -65,11 +65,11 @@
     }
 
     .prev {
-      .icon-arrow-left-thin();
+      .icon-arrow-left();
     }
 
     .next {
-      .icon-arrow-right-thin();
+      .icon-arrow-right4();
     }
 
     .month-container {
