+++ modules/admin-ui/src/main/resources/web/admin/common/styles/api/ui/time/time-picker.less
@@ -92,13 +92,13 @@
 
       .prev {
         span {
-          .icon-arrow-down-thin();
+          .icon-arrow-down4();
         }
       }
 
       .next {
         span {
-          .icon-arrow-up-thin();
+          .icon-arrow-up2();
         }
       }
