+++ modules/admin-ui/src/main/resources/web/admin/common/styles/api/ui/treegrid/tree-grid.less
@@ -164,11 +164,11 @@
       }
 
       .toggle.expand {
-        .icon-arrow-right3();
+        .icon-arrow-right2();
       }
 
       .toggle.collapse {
-        .icon-arrow-right3();
+        .icon-arrow-right2();
         &:before {
 
           .rotate(45deg);
