+++ modules/admin-ui/src/main/resources/web/admin/common/styles/apps/launcher/launcher.less
@@ -251,7 +251,7 @@ html, body {
     }
 
     .login-button {
-      .icon-arrow-right6;
+      .icon-arrow-right3;
       background-color: transparent;
       position: absolute;
       right: 0px;
