+++ modules/admin-ui/src/main/resources/web/admin/common/styles/styles.less
@@ -58,7 +58,7 @@
 @import "api/ui/time/date-time-picker";
 @import "api/ui/treegrid/tree-grid";
 @import "api/ui/drag-helper";
-@import "api/ui/image/focal-editor";
+@import "api/ui/image/image-editor";
 @import "api/ui/image/image-canvas";
 @import "api/aggregation/aggregation-group-view";
 @import "api/aggregation/aggregation-view";
